backend/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
pydantic==2.5.1
python-dotenv==1.0.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
pyjwt==2.10.1
torch==2.5.1
numpy==1.26.4
requests==2.32.0
